Move instruction wrapper types into a new file for discoverability
